refacto: remove inertie_thermique.py file for clearer infrastructure of the codebase
